+++ dimcli/tests/_test_converters.py
@@ -12,7 +12,7 @@ import configparser
 from .. import *
 from ..core.auth import USER_CONFIG_FILE_PATH
 from ..utils.dim_utils import dimensions_url
-from ..shortcuts import *
+from ..utils import *
 
 from ..core.converters import *