+++ dimcli/tests/test_dataframes.py
@@ -14,8 +14,7 @@ import configparser
 
 from .. import *
 from ..core.auth import USER_CONFIG_FILE_PATH
-from ..utils.dim_utils import dimensions_url
-from ..shortcuts import *
+from ..utils import *
 
 
 class TestDataframes(unittest.TestCase):