+++ dimcli/tests/test_grammar.py
@@ -13,8 +13,7 @@ import configparser
 
 from .. import *
 from ..core.auth import USER_CONFIG_FILE_PATH
-from ..utils.dim_utils import dimensions_url
-from ..shortcuts import *
+from ..utils import *
 
 
 class TestOne(unittest.TestCase):